# api/openapi.yaml
openapi: 3.1.0
info:
  title: SleekFlow TODO API
  version: 1.0.0
  description: Workspace-scoped collaborative TODO API. Errors use RFC 9457 Problem Details.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
tags:
  - name: Authentication
  - name: Workspaces
  - name: TODOs
paths:
  /auth/register:
    post:
      tags: [Authentication]
      operationId: register
      description: Create a new user account with an initial workspace, then sign in immediately with a session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterInput" }
      responses:
        "201":
          description: Registered and signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      description: Authenticate with email and password, establishing a session cookie and returning a CSRF token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      description: Revoke the current session and clear the session cookie.
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses:
        "204": { description: Current session revoked }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      description: Return the authenticated user's profile, a fresh CSRF token, and their accessible workspaces.
      responses:
        "200":
          description: Current authenticated user and CSRF token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUserResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /workspaces:
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      description: List all workspaces the current user can access, including their role in each.
      responses:
        "200":
          description: Workspaces visible to the current user
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/Workspace" },
                    }
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      description: Create a new workspace with the current user as owner.
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceInput" }
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "422": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    patch:
      tags: [Workspaces]
      operationId: updateWorkspace
      description: Update a workspace's name or timezone. Requires the owner role.
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceInput" }
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/members:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [Workspaces]
      operationId: listWorkspaceMembers
      description: List all members of a workspace with their roles.
      responses:
        "200":
          description: Workspace members
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/WorkspaceMember" },
                    }
    post:
      tags: [Workspaces]
      operationId: addWorkspaceMember
      description: Add an existing user to a workspace as an editor or viewer.
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/EditableRole" }
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceMember" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/members/{userId}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - {
          name: userId,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    patch:
      tags: [Workspaces]
      operationId: updateWorkspaceMember
      description: Change a workspace member's role between editor and viewer.
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/EditableRole" }
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceMember" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Workspaces]
      operationId: removeWorkspaceMember
      description: Remove a member from a workspace. The last owner cannot be removed.
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      responses:
        "204": { description: Member removed }
        "409": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/todos:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [TODOs]
      operationId: listTodos
      description: List TODOs in a workspace with filtering, sorting, and keyset pagination via cursor.
      parameters:
        - {
            name: status,
            in: query,
            explode: false,
            schema:
              {
                type: array,
                items: { $ref: "#/components/schemas/TodoStatus" },
              },
          }
        - {
            name: priority,
            in: query,
            explode: false,
            schema:
              {
                type: array,
                items: { $ref: "#/components/schemas/TodoPriority" },
              },
          }
        - {
            name: dueFrom,
            in: query,
            schema: { type: string, format: date-time },
          }
        - {
            name: dueTo,
            in: query,
            schema: { type: string, format: date-time },
          }
        - {
            name: dependencyState,
            in: query,
            schema: { type: string, enum: [blocked, unblocked] },
          }
        - {
            name: sort,
            in: query,
            schema:
              {
                type: string,
                enum: [dueAt, priority, status, name],
                default: dueAt,
              },
          }
        - {
            name: direction,
            in: query,
            schema: { type: string, enum: [asc, desc], default: asc },
          }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 50 },
          }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: Keyset-paginated TODOs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TodoList" }
        "422": { $ref: "#/components/responses/Problem" }
    post:
      tags: [TODOs]
      operationId: createTodo
      description: Create a TODO in a workspace, optionally with recurrence and initial dependencies.
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTodoInput" }
      responses:
        "201":
          description: Created TODO
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Todo" }
        "422": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/todos/{todoId}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/TodoId" }
    get:
      tags: [TODOs]
      operationId: getTodo
      description: Fetch a single TODO with its dependencies and current version (ETag).
      responses:
        "200":
          description: TODO details
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Todo" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [TODOs]
      operationId: updateTodo
      description: >-
        Update a TODO with optimistic concurrency (If-Match). Completing a
        recurring TODO generates the next occurrence in the series.
      parameters:
        - { $ref: "#/components/parameters/CsrfToken" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTodoInput" }
      responses:
        "200":
          description: Updated TODO and optionally generated recurrence occurrence
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TodoMutationResponse" }
        "409": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [TODOs]
      operationId: deleteTodo
      description: Soft-delete a TODO with optimistic concurrency (If-Match). Fails if other TODOs depend on it.
      parameters:
        - { $ref: "#/components/parameters/CsrfToken" }
        - { $ref: "#/components/parameters/IfMatch" }
      responses:
        "204": { description: Soft deleted }
        "409": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/todos/{todoId}/dependencies:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/TodoId" }
    post:
      tags: [TODOs]
      operationId: addTodoDependency
      description: Make a TODO depend on another TODO. Rejects cycles and self-dependencies.
      parameters:
        - { $ref: "#/components/parameters/CsrfToken" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dependsOnId]
              properties:
                dependsOnId: { type: string, format: uuid }
      responses:
        "200":
          description: Dependency added
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Todo" }
        "409": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/todos/{todoId}/dependencies/{dependsOnId}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
      - { $ref: "#/components/parameters/TodoId" }
      - {
          name: dependsOnId,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    delete:
      tags: [TODOs]
      operationId: removeTodoDependency
      description: Remove a dependency between two TODOs, unblocking the dependent TODO if applicable.
      parameters:
        - { $ref: "#/components/parameters/CsrfToken" }
        - { $ref: "#/components/parameters/IfMatch" }
      responses:
        "200":
          description: Dependency removed
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Todo" }
        "412": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/events:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [TODOs]
      operationId: streamWorkspaceEvents
      description: Server-sent events stream of committed workspace changes (created, updated, deleted TODOs) for real-time updates.
      responses:
        "200":
          description: Server-sent event stream of committed workspace changes
          content:
            text/event-stream:
              schema:
                type: string
                examples:
                  [
                    'event: todo.updated\ndata: {"eventId":"...","todoId":"...","version":2}\n\n',
                  ]
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-todo_session
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    TodoId:
      name: todoId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong or weak ETag containing the current integer version.
      schema: { type: string, examples: ['"3"'] }
  headers:
    ETag:
      description: Current TODO version
      schema: { type: string, examples: ['"3"'] }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        errors:
          type: object
          additionalProperties: { type: array, items: { type: string } }
        affectedTodos:
          type: array
          items: { $ref: "#/components/schemas/AffectedTodo" }
    User:
      type: object
      required: [id, email, displayName]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
    RegisterInput:
      type: object
      additionalProperties: false
      required: [email, password, displayName, workspaceName, timezone]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 128 }
        displayName: { type: string, minLength: 1, maxLength: 120 }
        workspaceName: { type: string, minLength: 1, maxLength: 120 }
        timezone: { type: string, examples: [Asia/Kuala_Lumpur] }
    LoginInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 1, maxLength: 128 }
    AuthResponse:
      type: object
      required: [user, csrfToken, workspaces]
      properties:
        user: { $ref: "#/components/schemas/User" }
        csrfToken: { type: string }
        workspaces:
          { type: array, items: { $ref: "#/components/schemas/Workspace" } }
    CurrentUserResponse:
      allOf:
        - { $ref: "#/components/schemas/AuthResponse" }
    WorkspaceRole:
      type: string
      enum: [owner, editor, viewer]
    EditableRole:
      type: string
      enum: [editor, viewer]
    Workspace:
      type: object
      required: [id, name, timezone, role, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        timezone: { type: string }
        role: { $ref: "#/components/schemas/WorkspaceRole" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    WorkspaceInput:
      type: object
      additionalProperties: false
      required: [name, timezone]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        timezone: { type: string }
    WorkspaceMember:
      type: object
      required: [userId, email, displayName, role, joinedAt]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { $ref: "#/components/schemas/WorkspaceRole" }
        joinedAt: { type: string, format: date-time }
    TodoStatus:
      type: string
      enum: [NotStarted, InProgress, Completed, Archived]
    TodoPriority:
      type: string
      enum: [Low, Medium, High]
    RecurrenceInput:
      type: object
      additionalProperties: false
      required: [intervalCount, intervalUnit]
      properties:
        intervalCount: { type: integer, minimum: 1, maximum: 365 }
        intervalUnit: { type: string, enum: [day, week, month] }
    Recurrence:
      allOf:
        - { $ref: "#/components/schemas/RecurrenceInput" }
        - type: object
          required: [seriesId, sequence, anchorLocal]
          properties:
            seriesId: { type: string, format: uuid }
            sequence: { type: integer, minimum: 0 }
            anchorLocal: { type: string }
    TodoDependency:
      type: object
      required: [id, name, status, completed]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { $ref: "#/components/schemas/TodoStatus" }
        completed: { type: boolean }
    Todo:
      type: object
      required:
        [
          id,
          workspaceId,
          name,
          description,
          dueAt,
          status,
          priority,
          version,
          blocked,
          blockingDependencyIds,
          dependencies,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        workspaceId: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        dueAt: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/TodoStatus" }
        priority: { $ref: "#/components/schemas/TodoPriority" }
        version: { type: integer, minimum: 1 }
        blocked: { type: boolean }
        blockingDependencyIds:
          { type: array, items: { type: string, format: uuid } }
        dependencies:
          {
            type: array,
            items: { $ref: "#/components/schemas/TodoDependency" },
          }
        recurrence:
          {
            oneOf:
              [{ $ref: "#/components/schemas/Recurrence" }, { type: "null" }],
          }
        completedAt:
          { oneOf: [{ type: string, format: date-time }, { type: "null" }] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateTodoInput:
      type: object
      additionalProperties: false
      required: [name, description, dueAt, status, priority]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 10000 }
        dueAt: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/TodoStatus" }
        priority: { $ref: "#/components/schemas/TodoPriority" }
        recurrence:
          {
            oneOf:
              [
                { $ref: "#/components/schemas/RecurrenceInput" },
                { type: "null" },
              ],
          }
        dependencyIds:
          {
            type: array,
            maxItems: 50,
            uniqueItems: true,
            items: { type: string, format: uuid },
          }
    UpdateTodoInput:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 10000 }
        dueAt: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/TodoStatus" }
        priority: { $ref: "#/components/schemas/TodoPriority" }
        cascadeDependents: { type: boolean, default: false }
      minProperties: 1
    TodoMutationResponse:
      type: object
      required: [todo]
      properties:
        todo: { $ref: "#/components/schemas/Todo" }
        generatedOccurrenceId:
          { oneOf: [{ type: string, format: uuid }, { type: "null" }] }
    TodoList:
      type: object
      required: [items, hasMore]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Todo" } }
        nextCursor: { oneOf: [{ type: string }, { type: "null" }] }
        hasMore: { type: boolean }
    AffectedTodo:
      type: object
      required: [id, name, status]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { $ref: "#/components/schemas/TodoStatus" }